+++ anx/provider/loadbalancer/reconciliation/reconcile_servers.go
@@ -11,16 +11,16 @@ import (
 
 const serverResourceTypeIdentifier = "01f321a4875446409d7d8469503a905f"
 
-func (r *reconciliation) filterServers(allServers []*lbaasv1.Server) ([]*lbaasv1.Server, error) {
+func (r *stateRetrieverImpl) filterServers(lbID string, allServers []*lbaasv1.Server) ([]*lbaasv1.Server, error) {
 	ret := make([]*lbaasv1.Server, 0, len(allServers))
 
 	for _, server := range allServers {
-		idx, err := compare.Search(lbaasv1.Backend{Identifier: server.Backend.Identifier}, r.backends, "Identifier")
+		idx, err := compare.Search(lbaasv1.Backend{Identifier: server.Backend.Identifier}, r.loadBalancers[lbID].state.backends, "Identifier")
 		if err != nil {
 			return nil, fmt.Errorf("error checking if Server belongs to one of our frontends: %w", err)
 		} else if idx != -1 {
 			ret = append(ret, server)
-			r.sortObjectIntoStateArray(server)
+			r.sortObjectIntoStateArray(lbID, server)
 		}
 	}
 
@@ -51,10 +51,10 @@ func (r *reconciliation) reconcileServers() (toCreate, toDestroy []types.Object,
 	}
 
 	toCreate = make([]types.Object, 0, len(targetServers))
-	toDestroy = make([]types.Object, 0, len(r.servers))
+	toDestroy = make([]types.Object, 0, len(r.remoteStateSnapshot.servers))
 
 	err = compare.Reconcile(
-		targetServers, r.servers,
+		targetServers, r.remoteStateSnapshot.servers,
 		&toCreate, &toDestroy,
 		"Name", "IP", "Port", "Check", "Backend.Identifier",
 	)